- Added multi-viewport story samples - File naming consistency
